pyproject.toml
[tool.poetry]
name = "molecule_dataset"
version = "0.1.0"
description = ""
authors = ["Kobi Felton <kobi.c.f@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
lorem = "^0.1.1"
pandas = "^1.1.1"
pillow = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
